instantiateSTAT: only drop AxisValues; always keep DesignAxes
when pinning an axis (ie. removing from fvar), we still need to keep the STAT's DesignAxis
as well as the single AxisValue table along that design axis that describes the
position of the new instance within the family's stylistic attributes.
We only match exact axis values, similarly to what we do with fvar's
named instances.